Make very clear that spamshield must be enabled
Change README.md to make it very clear that this extension relies on spamshield being enabled. The behavior, when spamshield is not enabled, is irritating for integrators: The form can be sent regardless of the captcha being solved or not. No error is logged. See e.g. two users who struggled with this in https://github.com/einpraegsam/powermailrecaptcha/issues/14